refactor buttons
