+++ apps/src/codebridge/ControlButtons/control-buttons.module.scss
@@ -1,0 +1,4 @@
+.controlButtonsContainer {
+  grid-area: control-buttons;
+  text-align: center;
+}
